fix+feat(underhell): GetDamage return type, DropWeapon, Throw_Nade, laser beam, exact SMG/minigun fire rate
- Fix C2555: CUHMeleeWeapon::GetDamage / CUHGunWeapon::GetDamage now return
  float to match CBaseEntity::GetDamage().
- DropWeapon command + m_bDisableWeaponDrop flag + Disable/EnableDropWeapon
  inputs (decode sub_101F11D0): un-sight, refuse in vehicle / sprinting /
  melee, toss forward at 300.
- Throw_Nade command (decode sub_101ED130): un-sight then start the grenade
  weapon's throw.
- Client laser sight (uh_laser.cpp): per-frame beam + dot from the eye along
  aim when m_bLaserToggleState is set (sprites/laserbeam.vmt + laserdot.vmt).
- Fire rates from serveror.dll: SMG mp5/mp5_eod/mp7 + BFG minigun = 0.075 s
  (GetFireRate vtable slot 277 -> sub_102801F0), G36K = 0.1 s confirmed.

+++ game/server/underhell/uh_weapons.h
@@ -42,7 +42,7 @@ class CUHMeleeWeapon : public CBaseHLBludgeonWeapon
 public:
 	virtual void	PrimaryAttack( void );	// drains StaminaToDrain, then swings
 	virtual float	GetDamageForActivity( Activity hitActivity ) { return m_flMeleeDamage; }
-	virtual int		GetDamage( void ) { return (int)m_flMeleeDamage; }
+	virtual float	GetDamage( void ) { return m_flMeleeDamage; }
 	virtual float	GetRange( void ) { return GetWpnData().m_flMeleeRange; }
 	virtual float	GetFireRate( void ) { return GetWpnData().m_flMeleeRoF; }
 
@@ -61,7 +61,7 @@ class CUHGunWeapon : public CBaseHLCombatWeapon
 
 public:
 	virtual void	PrimaryAttack( void );
-	virtual int		GetDamage( void ) { return m_iDamage; }
+	virtual float	GetDamage( void ) { return (float)m_iDamage; }
 	virtual float	GetFireRate( void ) { return m_flFireRate; }
 	virtual const Vector &GetBulletSpread( void );
 	virtual void	AddViewKick( void );
